feat: add shader parameters ui on viewer

+++ viewer/include/Shader.hpp
@@ -17,7 +17,7 @@ public:
     void unbind() const;
     void reload();
 
-    void setFloat(const std::string& name, int value) const;
+    void setFloat(const std::string& name, float value) const;
     void setVec3(const std::string& name, const Eigen::Vector3f& value) const;
     void setMat4(const std::string& name, const Eigen::Matrix4f& value) const;
